Delete contents in brackets

+++ rm_hw/src/hardware_interface/can_bus.cpp
@@ -239,9 +239,7 @@ void CanBus::read(ros::Time time)
     else if (data_ptr_.id2tof_data_->find(frame.can_id) != data_ptr_.id2tof_data_->end())
     {
       TofData& tof_data = data_ptr_.id2tof_data_->find(frame.can_id)->second;
-      tof_data.distance =
-          ((uint32_t)(((uint32_t)frame.data[2]) << 16) | ((uint32_t)frame.data[1] << 8) | (uint32_t)frame.data[0]) /
-          1000.;
+      tof_data.distance = (((frame.data[2]) << 16) | (frame.data[1] << 8) | frame.data[0]) / 1000.;
       tof_data.dis_status = frame.data[3];
       tof_data.signal_strength = ((int16_t)((frame.data[5]) << 8) | frame.data[4]);
       continue;
